Extract pasted-Markdown parsing into a markdown helper
parsePastedMarkdown(text) is the single export; the scoped Marked instance,
tag-lexeme escape, and element-taxonomy classification move with it, keeping
Clipboard to its OO surface.

Also drop the redundant ...marked.defaults spread: new Marked() already
starts from fresh defaults (gfm on, tables parsing), and snapshotting the
global singleton's mutable config at module load would let a host bundle's
marked.use() customizations leak into the paste flow.
